# Batch profile requests to speed up fetching