This speeds up loading the profile requests by batching the requests together and using `getProfiles` instead of `getProfile`.

## Comparison

On the audit log:

| Metric | Before | After |
| --- | --- | --- |
| Nuxt tools: Page load time | 12.5s | 1.7s |
| Chrome finish time | 14.8s | 5s |

## In summary

![](https://media.tenor.com/yRrrJYgNXfUAAAAd/the-dog.gif)